+++ sdk/src/org.graalvm.nativebridge/src/org/graalvm/nativebridge/GenerateNativeToHotSpotBridge.java
@@ -57,10 +57,8 @@ import java.lang.annotation.Target;
 public @interface GenerateNativeToHotSpotBridge {
 
     /**
-     * The native bridge configuration. The returned class must have an accessible static
-     * {@code getInstance()} method returning a {@link JNIConfig} instance. The returned
-     * {@link JNIConfig} instance is used for marshallers' lookup.
+     * Specifies the factory class that provides the configuration for the generated bridge, such as
+     * marshallers.
      */
-    Class<?> jniConfig();
-
+    Class<?> factory();
 }