+++ sdk/src/org.graalvm.nativebridge/src/org/graalvm/nativebridge/StackTraceElementMarshaller.java
@@ -50,7 +50,7 @@ final class StackTraceElementMarshaller implements BinaryMarshaller<StackTraceEl
     private static final int STACK_TRACE_ELEMENT_SIZE_ESTIMATE = 100;
 
     @Override
-    public StackTraceElement read(BinaryInput in) {
+    public StackTraceElement read(Isolate<?> isolate, BinaryInput in) {
         String className = in.readUTF();
         String methodName = in.readUTF();
         String fileName = in.readUTF();